backend-pjt/requirements.txt
Django==5.1.2
psycopg2-binary==2.9.10
djangorestframework==3.15.2
gunicorn==23.0.0
dj-rest-auth==6.0.0
django-allauth==65.1.0
djangorestframework-simplejwt==5.3.1
requests==2.32.3
django-health-check==3.18.3
django-cors-headers==4.5.0
openai==1.52.2
python-dotenv==1.0.1
langchain==0.3.4
pgvector==0.3.6
elasticsearch==8.17.1
django-stubs==5.1.0
black==24.10.0
pre-commit==4.0.1